+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "troncos"
-version = "2.2.1"
+version = "3.0.0"
 description = "Collection of Python logging and tracing tools"
 authors = [
     "Guðmundur Björn Birkisson <gudmundur.birkisson@oda.com>",
